CI Fix

+++ src/Application/src/Application.cpp
@@ -97,7 +97,7 @@ void Application::handleEvents(const ::sf::Event& event) {
         box->updateText(event.text.unicode);
     } else if (event.type == ::sf::Event::Resized) {
         ::sf::View newView = _window->getDefaultView();
-        newView.setSize(event.size.width, event.size.height);
+        newView.setSize((float)event.size.width, (float)event.size.height);
 
         _window->setView(newView);
     }
